pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "serious-news-agent"
version = "0.1.0"
description = "智能新闻助理系统 - 面向科技公司高管的个性化新闻流"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Development Team"}
]
keywords = ["news", "agent", "ai", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
]

dependencies = [
    # Web 框架
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",

    # 数据验证
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # 数据库
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.0",
    "aiosqlite>=0.19.0",  # SQLite 异步驱动
    "asyncpg>=0.29.0",  # PostgreSQL 异步驱动（可选）

    # HTTP 客户端
    "httpx>=0.25.0",

    # 任务调度
    "apscheduler>=3.10.0",

    # 环境配置
    "python-dotenv>=1.0.0",

    # 日志
    "loguru>=0.7.2",

    # Result 类型
    "returns>=0.23.0",

    # 文本处理 (去重功能)
    "scikit-learn>=1.8.0",

    # 监控
    "prometheus-client>=0.19.0",

    # 认证
    "bcrypt>=5.0.0",
    "PyJWT>=2.11.0",

    # Agent 框架 (需要从 GitHub 安装)
    # "nanobot-ai @ git+https://github.com/HKUDS/nanobot.git",
]

[project.optional-dependencies]
dev = [
    # 测试
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",  # 用于测试 FastAPI

    # 代码质量
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",

]

[project.urls]
Homepage = "https://github.com/your-org/serious-news-agent"
Repository = "https://github.com/your-org/serious-news-agent"

[project.scripts]
serious-news = "src.main:main"

# Black 配置
[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # 默认排除
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  # 自定义排除
  | \.kiro
  | \.claude
)/
'''

# Ruff 配置
[tool.ruff]
line-length = 100
target-version = "py311"

select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]

ignore = [
    "E501",  # 行长度由 Black 处理
    "B008",  # 函数调用中的默认参数（FastAPI 依赖）
    "B904",  # raise from (可能太严格）
    "ARG001",  # 未使用的参数（FastAPI 路由参数）
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # 允许未使用的导入
"tests/**/*.py" = ["ARG", "S101"]  # 测试中的未使用参数和 assert

[tool.ruff.isort]
known-first-party = ["src"]

# Pytest 配置
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--strict-markers",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

# Coverage 配置
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Mypy 配置
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "nanobot.*",
    "apscheduler.*",
    "returns.*",
    "sklearn.*",
]
ignore_missing_imports = true
